MICROSOFT-FOUNDRY/MICROSOFT-FOUNDRY-SDK/MCP-TOOL/requirements.txt
python-dotenv
azure-identity
uvicorn

# use the command 
# python -m venv labenv
# ./labenv/bin/Activate.ps1
# pip install -r requirements.txt --pre azure-ai-projects mcp